refactor : bascket-test 도메인 분리
